better commandline support

+++ src/types.h
@@ -6,7 +6,7 @@ typedef unsigned char u8;
 typedef unsigned short u16; 
 typedef signed short s16; 
 
-typedef unsigned long u32; 
-typedef signed long s32; 
+typedef unsigned int u32; 
+typedef signed int s32; 
 
 #endif
